move config.h to configuration.h and generate config.h instead of autoconfig.h
